LPS-146932 Rename Routine filename
