+++ modules/apps/site-initializer/site-initializer-testray/extra/remote-app/src/TestrayRouter.tsx
@@ -27,7 +27,7 @@ import ProjectOutlet from './pages/Project/ProjectOutlet';
 import Requirements from './pages/Project/Requirements';
 import AddRequirements from './pages/Project/Requirements/AddRequirements';
 import Requirement from './pages/Project/Requirements/Requirement';
-import Routine from './pages/Project/Routines';
+import Routines from './pages/Project/Routines';
 import Build from './pages/Project/Routines/Builds/Build';
 import BuildOutlet from './pages/Project/Routines/Builds/BuildOutlet';
 import CaseResult from './pages/Project/Routines/Builds/Inner/CaseResult';
@@ -38,9 +38,9 @@ import Components from './pages/Project/Routines/Builds/Inner/Components';
 import Results from './pages/Project/Routines/Builds/Inner/Results';
 import Runs from './pages/Project/Routines/Builds/Inner/Run';
 import Teams from './pages/Project/Routines/Builds/Inner/Teams';
+import Routine from './pages/Project/Routines/Routine';
 import RoutineArchived from './pages/Project/Routines/RoutineArchived';
 import RoutineOutlet from './pages/Project/Routines/RoutineOutlet';
-import Routines from './pages/Project/Routines/Routines';
 import Suites from './pages/Project/Suites';
 import Suite from './pages/Project/Suites/Suite';
 import Testflow from './pages/Testflow';
